+++ typecats/__about__.py
@@ -1,4 +1,4 @@
 """typecats"""
-__version__ = "1.5.0"
+__version__ = "1.6.0"
 __author__ = "Peter Gaultney"
 __author_email__ = "pgaultney@xoi.io"